update the train_pipeline
